requirements.txt
matplotlib==3.4.1
networkx==2.5.1
tqdm==4.60.0
pyYAML==5.4.1
pandas==1.2.4